requirements.txt
openai==0.27.2
torch==2.0.0
python-dotenv==1.0.0
google-auth-httplib2==0.1.0
httplib2==0.22.0
google-api-python-client==2.86.0